+++ src/shared/task/linkFile.ts
@@ -1,4 +1,4 @@
-import { normalizeLinkedFileIds } from './linkedFiles'
+import { normalizeLinkedFileIds } from './linkedFiles.ts'
 
 /** Merge a file id into task.linkedFileIds (deduped). */
 export function mergeLinkedFileId(